Find Phone Using Faster RCNN/requirements.txt
opencv-python 
numpy
torch 
torchvision
tqdm
scikit-learn
Pillow
collection
datetime
pycocotools

